+++ src/main/java/io/boomerang/Application.java
@@ -1,4 +1,4 @@
-package net.boomerangplatform;
+package io.boomerang;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -6,8 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import net.boomerangplatform.migration.MongockRunner;
+import io.boomerang.migration.MongockRunner;
 
 @SpringBootApplication
 public class Application implements CommandLineRunner {
