+++ src/main/java/io/boomerang/migration/MongockRunner.java
@@ -1,4 +1,4 @@
-package net.boomerangplatform.migration;
+package io.boomerang.migration;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
