+++ src/main/resources/application-flow.properties
@@ -1,1 +1,4 @@
 spring.data.mongodb.uri=mongodb://localhost:27017/boomerang
+
+
+flow.mongo.collection.prefix=flow
